requirements.txt
# Haggle Service Marketplace - Phase 1 Dependencies

# FastAPI framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Async HTTP client for Grok API calls
httpx>=0.27.0

# Database
sqlalchemy>=2.0.30

# Pydantic for data validation
pydantic>=2.9.0

# Environment variables
python-dotenv>=1.0.0
